Album controller routes, make API calls to Spotify
This commit connects the album controller routes to Spotify api (1 & 2)
https://developer.spotify.com/assets/AuthG_AuthoriztionCode.png

The naming of this commit is not good, nor is the location of this
route. In fact, albums should not even be a table in the backend, these
routes should go through the accounts table.
Next steps: move to the accounts table, create the front end generate
button which performs the access_token refresh functionality